Improves simplifier
